added addition + refactor

+++ src/main/java/net/focik/hr/employee/api/EmployeeController.java
@@ -2,7 +2,9 @@ package net.focik.hr.employee.api;
 
 import lombok.AllArgsConstructor;
 import lombok.extern.log4j.Log4j2;
-import net.focik.hr.employee.api.dto.*;
+import net.focik.hr.employee.api.dto.EmployeeDto;
+import net.focik.hr.employee.api.dto.EmployeeTypeDto;
+import net.focik.hr.employee.api.dto.EmployeeWorktimeDto;
 import net.focik.hr.employee.api.mapper.ApiEmployeeMapper;
 import net.focik.hr.employee.domain.Employee;
 import net.focik.hr.employee.domain.port.primary.AddNewEmployeeUseCase;
@@ -11,10 +13,9 @@ import net.focik.hr.employee.domain.port.primary.UpdateEmployeUseCase;
 import net.focik.hr.employee.domain.share.EmployeeType;
 import net.focik.hr.employee.domain.share.EmploymentStatus;
 import net.focik.hr.employee.domain.share.WorkTime;
-import net.focik.hr.employee.domain.utils.PrivilegeHelper;
 import net.focik.hr.utils.exceptions.AccessDeniedException;
 import net.focik.hr.utils.exceptions.HttpResponse;
-import org.modelmapper.ModelMapper;
+import net.focik.hr.utils.privileges.PrivilegeHelper;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -23,14 +24,14 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static net.focik.hr.employee.domain.utils.PrivilegeConstant.AUTHORITIES;
+import static net.focik.hr.utils.privileges.PrivilegeHelper.*;
 import static org.springframework.http.HttpStatus.OK;
 
 @Log4j2
 @RestController
 @AllArgsConstructor
 @RequestMapping("/api/employee")
-@CrossOrigin
+//@CrossOrigin
 class EmployeeController {
 
     private AddNewEmployeeUseCase addNewEmployeeUseCase;
@@ -42,11 +43,12 @@ class EmployeeController {
     public ResponseEntity<Integer> addEmployee(@RequestBody EmployeeDto employeeDto,
                                                @RequestHeader(name = AUTHORITIES, required = false) String[] roles) {
         log.info("Try add new employee.");
-        final List<String> accessRole = List.of("ROLE_ADMIN", "HR_EMPLOYEE_WRITE_ALL");
+        final List<String> accessRole = List.of(ROLE_ADMIN, HR_EMPLOYEE_WRITE_ALL);
 
-//        if (!PrivilegeHelper.checkAccess(List.of(roles), accessRole)) {
-//            throw new AccessDeniedException();
-//        }
+        if (PrivilegeHelper.dontHaveAccess(List.of(roles), accessRole)) {
+            throw new AccessDeniedException();
+        }
+
         Employee employee = mapper.toDomain(employeeDto);
         Integer result = addNewEmployeeUseCase.addEmployee(employee);
 
@@ -62,11 +64,11 @@ class EmployeeController {
     public ResponseEntity<EmployeeDto> updateEmployee(@RequestBody EmployeeDto employeeDto,
                                                       @RequestHeader(name = AUTHORITIES, required = false) String[] roles) {
         log.info("Try update employee.");
-        final List<String> accessRole = List.of("ROLE_ADMIN", "HR_EMPLOYEE_WRITE_ALL");
+        final List<String> accessRole = List.of(ROLE_ADMIN, HR_EMPLOYEE_WRITE_ALL);
 
-//        if (!PrivilegeHelper.checkAccess(List.of(roles), accessRole)) {
-//            throw new AccessDeniedException();
-//        }
+        if (PrivilegeHelper.dontHaveAccess(List.of(roles), accessRole)) {
+            throw new AccessDeniedException();
+        }
 
         Employee employee = updateEmployeUseCase.updateEmployee(mapper.toDomain(employeeDto));
         return new ResponseEntity<>(mapper.toDto(employee), OK);
@@ -75,18 +77,18 @@ class EmployeeController {
     @DeleteMapping("/{idEmployee}")
     public ResponseEntity<HttpResponse> deleteEmployee(@PathVariable int idEmployee,
                                                        @RequestHeader(name = AUTHORITIES, required = false) String[] roles) {
-        log.info("Try add new employee.");
-        final List<String> accessRole = List.of("ROLE_ADMIN", "HR_EMPLOYEE_DELETE_ALL");
+        log.info("Try delete employee with id: " + idEmployee);
+        final List<String> accessRole = List.of(ROLE_ADMIN, HR_EMPLOYEE_DELETE_ALL);
 
-//        if (!PrivilegeHelper.checkAccess(List.of(roles), accessRole)) {
-//            throw new AccessDeniedException();
-//        }
+        if (PrivilegeHelper.dontHaveAccess(List.of(roles), accessRole)) {
+            throw new AccessDeniedException();
+        }
 
         deleteEmployeeUseCase.deleteEmployeeById(idEmployee);
 
         log.info("Deleted employee with id = " + idEmployee);
 
-        return response(HttpStatus.NO_CONTENT, "Użytkownik usunięty.");
+        return response(HttpStatus.NO_CONTENT, "Pracownik usunięty.");
     }
 
     @GetMapping("/employeetype")
@@ -101,11 +103,11 @@ class EmployeeController {
     public ResponseEntity<HttpResponse> updateEmploymentStatus(@PathVariable int id, @RequestParam("status") EmploymentStatus employmentStatus,
                                                                @RequestHeader(name = AUTHORITIES, required = false) String[] roles) {
         log.info("Try update employment status.");
-        final List<String> accessRole = List.of("ROLE_ADMIN", "HR_EMPLOYEE_WRITE_ALL");
+        final List<String> accessRole = List.of(ROLE_ADMIN, HR_EMPLOYEE_WRITE_ALL);
 
-//        if (!PrivilegeHelper.checkAccess(List.of(roles), accessRole)) {
-//            throw new AccessDeniedException();
-//        }
+        if (PrivilegeHelper.dontHaveAccess(List.of(roles), accessRole)) {
+            throw new AccessDeniedException();
+        }
 
         updateEmployeUseCase.updateEmploymentStatus(id, employmentStatus);
         return response(HttpStatus.OK, "Zaaktualizowano status pracownika.");
